Introduce ParametricPSF base class
